+++ src/main/java/net/sf/jabref/gui/GenFieldsCustomizer.java
@@ -38,7 +38,7 @@ import javax.swing.JTextArea;
 import net.sf.jabref.Globals;
 import net.sf.jabref.gui.entryeditor.EntryEditorTabList;
 import net.sf.jabref.gui.help.HelpAction;
-import net.sf.jabref.gui.help.HelpFile;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.gui.keyboard.KeyBinding;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.labelpattern.LabelPatternUtil;
