+++ src/main/java/net/sf/jabref/gui/preftabs/NameFormatterTab.java
@@ -38,7 +38,7 @@ import javax.swing.table.TableModel;
 import net.sf.jabref.gui.IconTheme;
 import net.sf.jabref.gui.OSXCompatibleToolbar;
 import net.sf.jabref.gui.help.HelpAction;
-import net.sf.jabref.gui.help.HelpFile;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.layout.format.NameFormatter;
 import net.sf.jabref.preferences.JabRefPreferences;
