+++ src/main/java/net/sf/jabref/importer/fetcher/ACMPortalFetcher.java
@@ -43,7 +43,7 @@ import javax.swing.JRadioButton;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.gui.FetcherPreviewDialog;
-import net.sf.jabref.gui.help.HelpFile;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.importer.ImportInspector;
 import net.sf.jabref.importer.OutputPrinter;
 import net.sf.jabref.importer.fileformat.BibtexParser;
