+++ src/main/java/net/sf/jabref/importer/fetcher/DOItoBibTeXFetcher.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 import javax.swing.JPanel;
 
 import net.sf.jabref.Globals;
-import net.sf.jabref.gui.help.HelpFile;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.importer.ImportInspector;
 import net.sf.jabref.importer.OutputPrinter;
 import net.sf.jabref.importer.ParserResult;
