+++ src/main/java/net/sf/jabref/importer/fetcher/GVKFetcher.java
@@ -19,7 +19,7 @@ import java.util.Map;
 import javax.swing.JPanel;
 import javax.xml.parsers.ParserConfigurationException;
 
-import net.sf.jabref.gui.help.HelpFile;
+import net.sf.jabref.logic.help.HelpFile;
 import net.sf.jabref.importer.ImportInspector;
 import net.sf.jabref.importer.OutputPrinter;
 import net.sf.jabref.logic.l10n.Localization;
