+++ src/main/java/net/sf/jabref/importer/fetcher/SpringerFetcher.java
@@ -19,7 +19,7 @@ package net.sf.jabref.importer.fetcher;
     import javax.swing.JOptionPane;
     import javax.swing.JPanel;
 
-    import net.sf.jabref.gui.help.HelpFile;
+    import net.sf.jabref.logic.help.HelpFile;
     import net.sf.jabref.importer.ImportInspector;
     import net.sf.jabref.importer.OutputPrinter;
     import net.sf.jabref.importer.fileformat.JSONEntryParser;
